wat files concatenation
