+++ src/h264.wat
@@ -1,3 +1,7 @@
 (module
+    ;; ++++++++++++++++++++++++++++++++++++++++++++++++++++
+    ;; h264.wat
+    ;; ++++++++++++++++++++++++++++++++++++++++++++++++++++
+
     ;;
 )
